Use the Heroku platform API
The Heroku API exposed to Ruby plugins has been deprecated, and is
starting to suffer brownouts. As a result, the only way to continue
using this plugin is to fetch the needed information from the new API.
Unfortunately, the new API spreads out the three pieces of information
that we need across three separate API responses, which is... less than
ideal. So this plugin is slower than it used to be. But at least it
works again.